+++ src/frontends/lean/structure_instance.h
@@ -7,11 +7,17 @@ Author: Leonardo de Moura
 #pragma once
 #include "frontends/lean/parse_table.h"
 namespace lean {
-expr mk_structure_instance(name const & s, buffer<name> const & fns, buffer<expr> const & fvs);
-expr mk_structure_instance(expr const & src, buffer<name> const & fns, buffer<expr> const & fvs);
+expr mk_structure_instance(name const & s = {}, buffer<name> const & fns = {}, buffer<expr> const & fvs = {},
+                           buffer<expr> const & sources = {}, bool catchall = false);
 bool is_structure_instance(expr const & e);
-void get_structure_instance_info(expr const & e, name & struct_name, optional<expr> & source,
-                                 buffer<name> & field_names, buffer<expr> & field_values);
+struct structure_instance_info {
+    name m_struct_name; // empty if omitted
+    buffer<name> m_field_names;
+    buffer<expr> m_field_values;
+    buffer<expr> m_sources;
+    bool m_catchall; // "..." syntax: fill in placeholders for remaining fields
+};
+structure_instance_info get_structure_instance_info(expr const & e);
 void initialize_structure_instance();
 void finalize_structure_instance();
 }