+++ src/frontends/lean/tokens.cpp
@@ -54,6 +54,7 @@ static name const * g_assume_tk = nullptr;
 static name const * g_suppose_tk = nullptr;
 static name const * g_fun_tk = nullptr;
 static name const * g_match_tk = nullptr;
+static name const * g_dotdot_tk = nullptr;
 static name const * g_ellipsis_tk = nullptr;
 static name const * g_raw_tk = nullptr;
 static name const * g_true_tk = nullptr;
@@ -179,6 +180,7 @@ void initialize_tokens() {
     g_suppose_tk = new name{"suppose"};
     g_fun_tk = new name{"fun"};
     g_match_tk = new name{"match"};
+    g_dotdot_tk = new name{".."};
     g_ellipsis_tk = new name{"..."};
     g_raw_tk = new name{"raw"};
     g_true_tk = new name{"true"};
@@ -305,6 +307,7 @@ void finalize_tokens() {
     delete g_suppose_tk;
     delete g_fun_tk;
     delete g_match_tk;
+    delete g_dotdot_tk;
     delete g_ellipsis_tk;
     delete g_raw_tk;
     delete g_true_tk;
@@ -430,6 +433,7 @@ name const & get_assume_tk() { return *g_assume_tk; }
 name const & get_suppose_tk() { return *g_suppose_tk; }
 name const & get_fun_tk() { return *g_fun_tk; }
 name const & get_match_tk() { return *g_match_tk; }
+name const & get_dotdot_tk() { return *g_dotdot_tk; }
 name const & get_ellipsis_tk() { return *g_ellipsis_tk; }
 name const & get_raw_tk() { return *g_raw_tk; }
 name const & get_true_tk() { return *g_true_tk; }