+++ src/frontends/lean/tokens.h
@@ -56,6 +56,7 @@ name const & get_assume_tk();
 name const & get_suppose_tk();
 name const & get_fun_tk();
 name const & get_match_tk();
+name const & get_dotdot_tk();
 name const & get_ellipsis_tk();
 name const & get_raw_tk();
 name const & get_true_tk();