+++ doc/changes.md
@@ -25,8 +25,15 @@ master branch (aka work in progress branch)
   The new instance is more efficient than the one in mathlib because it doesn't
   convert the array into a list.
 
+
 - Add aliasing pattern syntax `id@pat`, which introduces the name `id` for the value matched by
   the pattern `pat`.
+
+- Add alternative syntax `{..., ..s}` for the structure update `{s with ...}`.
+  Multiple fallback sources can be given: `{..., ..s, ..t}` will fall back to
+  searching a field in `s`, then in `t`. The last component can also be `..`,
+  which will replace any missing fields with a placeholder.
+  The old notation will be removed in the future.
 
 *Changes*